move cut classes to a separate module space
