+++ prd.py
@@ -120,7 +120,6 @@ from typing import Text, List, Union
 import torch
 from torch import nn
 from torch.nn import functional as F
-import torchvision.transforms as T
 import torchvision.transforms.functional as TF
 from tqdm import tqdm
 sys.path.append(f'{root_path}/ResizeRight')
@@ -137,6 +136,7 @@ import urllib.request
 from os.path import exists
 
 from helpers.utils import fetch
+from cut_modules.make_cutouts import MakeCutoutsDango, MakeCutouts
 
 
 # Setting default values for everything, which can then be overridden by settings files.
@@ -1139,217 +1139,6 @@ def read_image_workaround(path):
     this incompatibility to avoid colour inversions."""
     im_tmp = cv2.imread(path)
     return cv2.cvtColor(im_tmp, cv2.COLOR_BGR2RGB)
-
-
-def sinc(x):
-    return torch.where(x != 0,
-                       torch.sin(math.pi * x) / (math.pi * x), x.new_ones([]))
-
-
-def lanczos(x, a):
-    cond = torch.logical_and(-a < x, x < a)
-    out = torch.where(cond, sinc(x) * sinc(x / a), x.new_zeros([]))
-    return out / out.sum()
-
-
-def ramp(ratio, width):
-    n = math.ceil(width / ratio + 1)
-    out = torch.empty([n])
-    cur = 0
-    for i in range(out.shape[0]):
-        out[i] = cur
-        cur += ratio
-    return torch.cat([-out[1:].flip([0]), out])[1:-1]
-
-
-def resample(input, size, align_corners=True):
-    n, c, h, w = input.shape
-    dh, dw = size
-
-    input = input.reshape([n * c, 1, h, w])
-
-    if dh < h:
-        kernel_h = lanczos(ramp(dh / h, 2), 2).to(input.device, input.dtype)
-        pad_h = (kernel_h.shape[0] - 1) // 2
-        input = F.pad(input, (0, 0, pad_h, pad_h), 'reflect')
-        input = F.conv2d(input, kernel_h[None, None, :, None])
-
-    if dw < w:
-        kernel_w = lanczos(ramp(dw / w, 2), 2).to(input.device, input.dtype)
-        pad_w = (kernel_w.shape[0] - 1) // 2
-        input = F.pad(input, (pad_w, pad_w, 0, 0), 'reflect')
-        input = F.conv2d(input, kernel_w[None, None, None, :])
-
-    input = input.reshape([n, c, h, w])
-    return F.interpolate(input,
-                         size,
-                         mode='bicubic',
-                         align_corners=align_corners)
-
-
-class MakeCutouts(nn.Module):
-    def __init__(self, cut_size, cutn, skip_augs=False):
-        super().__init__()
-        self.cut_size = cut_size
-        self.cutn = cutn
-        self.skip_augs = skip_augs
-        self.augs = T.Compose([
-            T.RandomHorizontalFlip(p=0.5),
-            T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-            T.RandomAffine(degrees=15, translate=(0.1, 0.1)),
-            T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-            T.RandomPerspective(distortion_scale=0.4, p=0.7),
-            T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-            T.RandomGrayscale(p=0.15),
-            T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-            # T.ColorJitter(brightness=0.1, contrast=0.1, saturation=0.1, hue=0.1),
-        ])
-
-    def forward(self, input):
-        input = T.Pad(input.shape[2] // 4, fill=0)(input)
-        sideY, sideX = input.shape[2:4]
-        max_size = min(sideX, sideY)
-
-        cutouts = []
-        for ch in range(self.cutn):
-            if ch > self.cutn - self.cutn // 4:
-                cutout = input.clone()
-            else:
-                size = int(max_size * torch.zeros(1, ).normal_(
-                    mean=.8, std=.3).clip(float(self.cut_size / max_size), 1.))
-                offsetx = torch.randint(0, abs(sideX - size + 1), ())
-                offsety = torch.randint(0, abs(sideY - size + 1), ())
-                cutout = input[:, :, offsety:offsety + size,
-                               offsetx:offsetx + size]
-
-            if not self.skip_augs:
-                cutout = self.augs(cutout)
-            cutouts.append(resample(cutout, (self.cut_size, self.cut_size)))
-            del cutout
-
-        cutouts = torch.cat(cutouts, dim=0)
-        return cutouts
-
-
-cutout_debug = False
-padargs = {}
-
-
-class MakeCutoutsDango(nn.Module):
-    def __init__(self,
-                 cut_size,
-                 Overview=4,
-                 InnerCrop=0,
-                 IC_Size_Pow=0.5,
-                 IC_Grey_P=0.2):
-        super().__init__()
-        self.cut_size = cut_size
-        self.Overview = Overview
-        self.InnerCrop = InnerCrop
-        self.IC_Size_Pow = IC_Size_Pow
-        self.IC_Grey_P = IC_Grey_P
-        if args.animation_mode == 'None':
-            self.augs = T.Compose([
-                T.RandomHorizontalFlip(p=0.5),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomAffine(degrees=10,
-                               translate=(0.05, 0.05),
-                               interpolation=T.InterpolationMode.BILINEAR),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomGrayscale(p=0.1),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.ColorJitter(brightness=0.1,
-                              contrast=0.1,
-                              saturation=0.1,
-                              hue=0.1),
-            ])
-        elif args.animation_mode == 'Video Input':
-            self.augs = T.Compose([
-                T.RandomHorizontalFlip(p=0.5),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomAffine(degrees=15, translate=(0.1, 0.1)),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomPerspective(distortion_scale=0.4, p=0.7),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomGrayscale(p=0.15),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                # T.ColorJitter(brightness=0.1, contrast=0.1, saturation=0.1, hue=0.1),
-            ])
-        elif args.animation_mode == '2D' or args.animation_mode == '3D':
-            self.augs = T.Compose([
-                T.RandomHorizontalFlip(p=0.4),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomAffine(degrees=10,
-                               translate=(0.05, 0.05),
-                               interpolation=T.InterpolationMode.BILINEAR),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.RandomGrayscale(p=0.1),
-                T.Lambda(lambda x: x + torch.randn_like(x) * 0.01),
-                T.ColorJitter(brightness=0.1,
-                              contrast=0.1,
-                              saturation=0.1,
-                              hue=0.3),
-            ])
-
-    def forward(self, input):
-        cutouts = []
-        gray = T.Grayscale(3)
-        sideY, sideX = input.shape[2:4]
-        max_size = min(sideX, sideY)
-        min_size = min(sideX, sideY, self.cut_size)
-        l_size = max(sideX, sideY)
-        output_shape = [1, 3, self.cut_size, self.cut_size]
-        output_shape_2 = [1, 3, self.cut_size + 2, self.cut_size + 2]
-        pad_input = F.pad(input,
-                          ((sideY - max_size) // 2, (sideY - max_size) // 2,
-                           (sideX - max_size) // 2, (sideX - max_size) // 2),
-                          **padargs)
-        cutout = resize(pad_input, out_shape=output_shape)
-
-        # create a list of 1 to 4 in random order, then do the matching overview cut
-        # This way even with less than 4 overview cuts, you still get a mix of all of them   
-        if self.Overview > 0:
-            if self.Overview <= 4:
-                li = [1, 1, 2, 3, 4] # give a slight edge to the normal, full color cut
-                ri = random.sample(li,self.Overview)
-                ri.sort()
-                for i in range(self.Overview):
-                    if ri[i] == 1:
-                        cutouts.append(cutout)
-                    if ri[i] == 2:
-                        cutouts.append(gray(cutout))
-                    if ri[i] == 3:
-                        cutouts.append(TF.hflip(cutout))
-                    if ri[i] == 4:
-                        cutouts.append(gray(TF.hflip(cutout)))
-            else:
-                cutout = resize(pad_input, out_shape=output_shape)
-                for _ in range(self.Overview):
-                    cutouts.append(cutout)
-
-            if cutout_debug:
-                TF.to_pil_image(cutouts[0].clamp(0, 1).squeeze(0)).save(
-                    "content/cutout_overview0.jpg", quality=99)
-
-        if self.InnerCrop > 0:
-            for i in range(self.InnerCrop):
-                size = int(
-                    torch.rand([])**self.IC_Size_Pow * (max_size - min_size) +
-                    min_size)
-                offsetx = torch.randint(0, sideX - size + 1, ())
-                offsety = torch.randint(0, sideY - size + 1, ())
-                cutout = input[:, :, offsety:offsety + size,
-                               offsetx:offsetx + size]
-                if i <= int(self.IC_Grey_P * self.InnerCrop):
-                    cutout = gray(cutout)
-                cutout = resize(cutout, out_shape=output_shape)
-                cutouts.append(cutout)
-            if cutout_debug:
-                TF.to_pil_image(cutouts[-1].clamp(0, 1).squeeze(0)).save(
-                    "content/cutout_InnerCrop.jpg", quality=99)
-        cutouts = torch.cat(cutouts)
-        if skip_augs is not True: cutouts = self.augs(cutouts)
-        return cutouts
 
 
 def spherical_dist_loss(x, y):
